MUST_REBASE [VPG]: drm/i915: hdmi: disable interlaced mode
To prune the interlaced mode from mode list, this patch
disables the interlaced mode on hdmi connector.

We are pruning the interlaced modes as we are not supporting interlaced
modes on HSW PV1. On long run we will move this pruning to user mode and
will revert this change. Hence this change is not for upstreaming only for
OTC android's local merge.

Change-Id: Ic55cffea1b017fa163682e25c255e1777c4c89f9
Signed-off-by: Ramalingam C <ramalingam.c@intel.com>
(cherry picked from commit dfe6b0045ddedc3fb99b039b838e033e55bd1c64)

Signed-off-by: Jon Bloomfield <jon.bloomfield@intel.com>